refactor: GpuServerViewRespoinse 타입 정의 추가
